Crate lz

lz

A command-line interface for various math/tech subjects. Based on the ladderz library.

Installation

To install the command-line interface, run the following command in your terminal:

cargo install --git https://github.com/rzmk/ladderz --branch main

Example

lz prealgebra is-factor 3 12

3 is a factor of 12.

You may view the help menu for a subject and a function by running the command with the -h or --help flag:

lz prealgebra is-factor -h

Learn more on GitHub.

Modules

dsa

prealgebra

Structs

Cli 🔒

Enums

Subjects 🔒

The subjects that can be used.

Functions

main 🔒