Enum lz::prealgebra::Prealgebra

pub enum Prealgebra {
    FactorPairs {
        n: u32,
        raw: bool,
    },
    Factors {
        n: u32,
        raw: bool,
    },
    MultiplesInRange {
        n: u32,
        lower_bound: u32,
        upper_bound: u32,
        raw: bool,
    },
    PrimesInRange {
        lower_bound: u32,
        upper_bound: u32,
        raw: bool,
    },
    PrimeFactorization {
        n: u32,
        raw: bool,
    },
    IsComposite {
        n: u32,
        raw: bool,
    },
    IsPrime {
        n: u32,
        raw: bool,
    },
    IsFactor {
        n: u32,
        m: u32,
        raw: bool,
    },
    IsMultiple {
        n: u32,
        m: u32,
        raw: bool,
    },
}

Variants

FactorPairs

Fields

n: u32

The positive integer to find factor pairs for.

raw: bool

Whether or not to return the raw output.

Finds all factor pairs for a positive integer.

Example

Input

lz prealgebra factor-pairs 12

Output

The factor pairs of 12 are {(1, 12), (2, 6), (3, 4)}.

Raw Output (use -r or --raw)

{(1, 12), (2, 6), (3, 4)}

Factors

Fields

n: u32

The positive integer to find factors for.

raw: bool

Whether or not to return the raw output.

Finds all factors for a positive integer.

Example

Input

lz prealgebra factors 12

Output

The factors of 12 are {1, 2, 3, 4, 6, 12}.

Raw Output (use -r or --raw)

{1, 2, 3, 4, 6, 12}

MultiplesInRange

Fields

n: u32

The positive integer to find multiples for.

lower_bound: u32

The lower bound of the range to find multiples in.

upper_bound: u32

The upper bound of the range to find multiples in.

raw: bool

Whether or not to return the raw output.

Finds all multiples of a positive integer in a given range.

Example

Input

lz prealgebra multiples-in-range 3 10

Output

The multiples of 3 in the range [1, 10] are {3, 6, 9}.

Raw Output (use -r or --raw)

{3, 6, 9}

PrimesInRange

Fields

lower_bound: u32

The lower bound of the range to find primes in.

upper_bound: u32

The upper bound of the range to find primes in.

raw: bool

Whether or not to return the raw output.

Finds all primes in a given range.

Example

Input

lz prealgebra primes-in-range 1 10

Output

The primes in the range [1, 10] are {2, 3, 7, 5}.

Raw Output (use -r or --raw)

{2, 3, 7, 5}

PrimeFactorization

Fields

n: u32

The positive integer to find the prime factorization of.

raw: bool

Whether or not to return the raw output.

Finds the prime factorization of a positive integer.

Example

Input

lz prealgebra prime-factorization 12

Output

The prime factorization of 12 is {3: 1, 2: 2}.

Raw Output (use -r or --raw)

{3: 1, 2: 2}

IsComposite

Fields

n: u32

The positive integer to determine if it is composite.

raw: bool

Whether or not to return the raw output.

Determines if a positive integer is composite.

Example

Input

lz prealgebra is-composite 12

Output

12 is composite.

Raw Output (use -r or --raw)

true

IsPrime

Fields

n: u32

The positive integer to determine if it is prime.

raw: bool

Whether or not to return the raw output.

Determines if a positive integer is prime.

Example

Input

lz prealgebra is-prime 12

Output

12 is not prime.

Raw Output (use -r or --raw)

false

IsFactor

Fields

n: u32

The positive integer to determine if it is a factor.

m: u32

The positive integer to determine if it is a multiple.

raw: bool

Whether or not to return the raw output.

Determines if a positive integer is a factor of another positive integer.

Example

Input

lz prealgebra is-factor 3 12

Output

3 is a factor of 12.

Raw Output (use -r or --raw)

true

IsMultiple

Fields

n: u32

The positive integer to determine if it is a multiple.

m: u32

The positive integer to determine if it is a factor.

raw: bool

Whether or not to return the raw output.

Determines if a positive integer is a multiple of another positive integer.

Example

Input

lz prealgebra is-multiple 12 3

Output

12 is a multiple of 3.

Raw Output (use -r or --raw)

true